+++ ta4j-core/src/main/java/org/ta4j/core/criteria/drawdown/MonteCarloMaximumDrawdownCriterion.java
@@ -33,7 +33,7 @@ import org.ta4j.core.Position;
 import org.ta4j.core.TradingRecord;
 import org.ta4j.core.analysis.CashFlow;
 import org.ta4j.core.criteria.AbstractAnalysisCriterion;
-import org.ta4j.core.criteria.helpers.Statistics;
+import org.ta4j.core.criteria.Statistics;
 import org.ta4j.core.num.Num;
 
 /**
@@ -50,7 +50,7 @@ public class MonteCarloMaximumDrawdownCriterion extends AbstractAnalysisCriterio
     private final int iterations;
     private final Integer pathBlocks;
     private final Supplier<RandomGenerator> randomSupplier;
-    private final Statistics statistic;
+    private final Statistics statistics;
 
     /**
      * Default constructor returning the 95th percentile.
@@ -68,13 +68,13 @@ public class MonteCarloMaximumDrawdownCriterion extends AbstractAnalysisCriterio
      * @param pathBlocks number of trades to include in each simulated path
      *                   ({@code null} = use the number of trades in the sample)
      * @param seed       random seed for reproducibility
-     * @param statistic  which summary statistic of the simulated drawdowns to
+     * @param statistics which summary statistic of the simulated drawdowns to
      *                   return
      *
      * @since 0.19
      */
-    public MonteCarloMaximumDrawdownCriterion(int iterations, Integer pathBlocks, long seed, Statistics statistic) {
-        this(iterations, pathBlocks, () -> new SplittableRandom(seed), statistic);
+    public MonteCarloMaximumDrawdownCriterion(int iterations, Integer pathBlocks, long seed, Statistics statistics) {
+        this(iterations, pathBlocks, () -> new SplittableRandom(seed), statistics);
     }
 
     /**
@@ -84,17 +84,17 @@ public class MonteCarloMaximumDrawdownCriterion extends AbstractAnalysisCriterio
      * @param pathBlocks     number of trades to include in each simulated path
      *                       ({@code null} = use the number of trades in the sample)
      * @param randomSupplier supplier of the random generator used for simulations
-     * @param statistic      which summary statistic of the simulated drawdowns to
+     * @param statistics     which summary statistics of the simulated drawdowns to
      *                       return
      *
      * @since 0.19
      */
     public MonteCarloMaximumDrawdownCriterion(int iterations, Integer pathBlocks,
-            Supplier<RandomGenerator> randomSupplier, Statistics statistic) {
+            Supplier<RandomGenerator> randomSupplier, Statistics statistics) {
         this.iterations = iterations;
         this.pathBlocks = pathBlocks;
         this.randomSupplier = randomSupplier;
-        this.statistic = statistic;
+        this.statistics = statistics;
     }
 
     /**
@@ -146,7 +146,7 @@ public class MonteCarloMaximumDrawdownCriterion extends AbstractAnalysisCriterio
             }
             maxDrawdowns[iteration] = maxDrawdown.doubleValue();
         }
-        var result = Statistics.calculate(maxDrawdowns, statistic);
+        var result = statistics.calculate(maxDrawdowns);
         return numFactory.numOf(result);
     }