+++ ta4j-core/src/test/java/org/ta4j/core/criteria/drawdown/MonteCarloMaximumDrawdownCriterionTest.java
@@ -32,7 +32,7 @@ import org.ta4j.core.BaseTradingRecord;
 import static org.ta4j.core.TestUtils.assertNumEquals;
 import org.ta4j.core.Trade;
 import org.ta4j.core.criteria.AbstractCriterionTest;
-import static org.ta4j.core.criteria.helpers.Statistics.*;
+import static org.ta4j.core.criteria.Statistics.*;
 import org.ta4j.core.mocks.MockBarSeriesBuilder;
 import org.ta4j.core.num.NumFactory;